feat: cache recommendations and optimize images
